% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{render_ind_report_shiny}
\alias{render_ind_report_shiny}
\title{Render shiny indicator report}
\usage{
render_ind_report_shiny(x, save_data = TRUE, input = "package", file_var)
}
\arguments{
\item{x}{}

\item{save_data}{}

\item{file_var}{}
}
\description{
Render shiny indicator report
}
