% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{render_ind_page_shiny}
\alias{render_ind_page_shiny}
\title{Render shiny indicator page}
\usage{
render_ind_page_shiny(x, input, file)
}
\arguments{
\item{x}{}

\item{file_var}{}
}
\description{
Render shiny indicator page
}
