% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{render_reg_report_shiny}
\alias{render_reg_report_shiny}
\title{Render shiny regression report}
\usage{
render_reg_report_shiny(
  stock_var,
  epus_var = "MAB",
  region_var = "Mid",
  remove_var = FALSE,
  lag_var = 0,
  save_var = TRUE,
  this_dir,
  file_var
)
}
\arguments{
\item{stock_var}{}

\item{epus_var}{}

\item{file_var}{}
}
\description{
Render shiny regression report
}
